Enhance OnRail methods and update related tests
- Added multiple overloads of the `OnRail` extension method in `UpshotExtensions.cs` for improved success and failure handling.
- Replaced an older `OnRail` method in `UpshotTypeExtensions.cs` with a more versatile version using predicates.
- Streamlined the `Map` method in `UpshotTypeExtensions.cs` to ensure consistent return types and improved failure handling.
- Updated the `StartRailWay_ShouldWrapOperationAndReturnUpshotMessageError` test in `UpshotTests.cs` to reflect changes in return types from `Upshot<int>` to `Upshot<double>`.

+++ tests/RiseOn.RailResult.Upshot.Tests/UpshotTests.cs
@@ -48,14 +48,14 @@ public class UpshotTests
         // Arrange
         var resultOperation = 100;
         var initialValue = 10;
-        
-        Func<int, Upshot<int>> failFun = value => Upshot<int>.Success(value / 10);
-        Func<int, Upshot<int>> successFun = value => Upshot<int>.Success(value * 10);
+
+        Func<int, Upshot<double>> failFun = value => Upshot<double>.Success(value / 10);
+        Func<int, Upshot<double>> successFun = value => Upshot<double>.Success(value * 10);
 
         // Act
         var result = initialValue
             .OnRail(x => x > 10,
-                failRail: failFun, 
+                failRail: failFun,
                 successRail: successFun);
     }
 }
